general cleanup, improve documentation
